Add config changes
